uart

+++ MOZA_SDK/1.0.0.6/MSVC2019-64/example/QtWidgetsApplication2.h
@@ -22,8 +22,8 @@ private:
     Ui::Widget* ui;
     int startAngle;//起始角度
     double angle;
-    float currentValue;//当前值
-    float currentValue2;
+    int currentValue;//当前值
+    int currentValue2;
     QTimer* timer;
     bool flag;//指针转动标志位
